# Fixes to filter out undesirable fences 