--- a/debugging/kernel-logger/kp_kernel_logger.cpp
+++ b/debugging/kernel-logger/kp_kernel_logger.cpp
@@ -97,7 +97,7 @@ extern "C" void kokkosp_begin_parallel_for(const char* name, const uint32_t devI
 	int level = kokkosp_print_region_stack();
 	kokkosp_print_region_stack_indent(level);
 	
-	printf(" %s\n", name);
+	printf("    %s\n", name);
 }
 
 extern "C" void kokkosp_end_parallel_for(const uint64_t kID) {
@@ -113,7 +113,7 @@ extern "C" void kokkosp_begin_parallel_scan(const char* name, const uint32_t dev
 	int level = kokkosp_print_region_stack();
 	kokkosp_print_region_stack_indent(level);
 	
-	printf(" %s\n", name);
+	printf("    %s\n", name);
 }
 
 extern "C" void kokkosp_end_parallel_scan(const uint64_t kID) {
@@ -129,7 +129,7 @@ extern "C" void kokkosp_begin_parallel_reduce(const char* name, const uint32_t d
 	int level = kokkosp_print_region_stack();
 	kokkosp_print_region_stack_indent(level);
 	
-	printf(" %s\n", name);
+	printf("    %s\n", name);
 }
 
 extern "C" void kokkosp_end_parallel_reduce(const uint64_t kID) {
@@ -145,7 +145,7 @@ extern "C" void kokkosp_begin_fence(const char* name, const uint32_t devID, uint
 	int level = kokkosp_print_region_stack();
 	kokkosp_print_region_stack_indent(level);
 	
-	printf(" %s\n", name);
+	printf("    %s\n", name);
 }
 
 extern "C" void kokkosp_end_fence(const uint64_t kID) {