--- a/profiling/nvprof-connector/kp_nvprof_connector.cpp
+++ b/profiling/nvprof-connector/kp_nvprof_connector.cpp
@@ -119,26 +119,26 @@ extern "C" void kokkosp_pop_profile_region() {
 
 namespace {
 struct Section {
-        std::string label;
-        nvtxRangeId_t id;
+  std::string label;
+  nvtxRangeId_t id;
 };
 std::vector<Section> kokkosp_sections;
 }  // namespace
 
 extern "C" void kokkosp_create_profile_section(const char* name,
                                                uint32_t* sID) {
-        *sID = kokkosp_sections.size();
-        kokkosp_sections.push_back(
-            {std::string(name), static_cast<nvtxRangeId_t>(-1)});
+  *sID = kokkosp_sections.size();
+  kokkosp_sections.push_back(
+    {std::string(name), static_cast<nvtxRangeId_t>(-1)});
 }
 
 extern "C" void kokkosp_start_profile_section(const uint32_t sID) {
-        auto& section = kokkosp_sections[sID];
-        section.id = nvtxRangeStartA(section.label.c_str());
+  auto& section = kokkosp_sections[sID];
+  section.id = nvtxRangeStartA(section.label.c_str());
 }
 
 extern "C" void kokkosp_stop_profile_section(const uint32_t sID) {
-        auto const& section = kokkosp_sections[sID];
-        nvtxRangeEnd(section.id);
+  auto const& section = kokkosp_sections[sID];
+  nvtxRangeEnd(section.id);
 }