--- a/profiling/roctx-connector/Makefile
+++ b/profiling/roctx-connector/Makefile
@@ -1,5 +1,5 @@
 CXX=g++
-CXXFLAGS=-O3 -g -I$(ROCM_PATH)/roctracer/include
+CXXFLAGS=-O3 -g -I$(ROCM_PATH)/include/roctracer
 LDFLAGS=-L$(ROCM_PATH)/lib
 LIBS=-lroctx64
 SHARED_CXXFLAGS=-shared -fPIC